+++ src/full/Agda/Syntax/Abstract/Views.hs
@@ -183,6 +183,7 @@ instance ExprLike Expr where
       DontCare e                 -> DontCare <$> recurse e
       PatternSyn{}               -> pure e0
       Macro{}                    -> pure e0
+      Qualified q e              -> Qualified q <$> recurse e
 
   foldExpr :: forall m. FoldExprFn m Expr
   foldExpr f e =
@@ -215,6 +216,7 @@ instance ExprLike Expr where
       QuoteTerm{}              -> m
       Unquote{}                -> m
       DontCare e               -> m `mappend` fold e
+      Qualified _ e            -> m `mappend` fold e
    where
      m = f e
      fold :: FoldExprRecFn m
@@ -254,6 +256,7 @@ instance ExprLike Expr where
       DontCare e                 -> f =<< DontCare <$> trav e
       PatternSyn{}               -> f e
       Macro{}                    -> f e
+      Qualified m e              -> f =<< Qualified m <$> trav e
 
 instance ExprLike a => ExprLike (Arg a)
 instance ExprLike a => ExprLike (Maybe a)