+++ src/full/Agda/Syntax/Parser/Tokens.hs
@@ -1,7 +1,8 @@
 {-# OPTIONS_GHC -Wunused-imports #-}
 
 module Agda.Syntax.Parser.Tokens
-    ( Token(..)
+    ( Token(.., TokQual)
+    , QualifiableToken(..), QualifiedToken(..)
     , Keyword(..)
     , layoutKeywords
     , Symbol(..)
@@ -67,17 +68,51 @@ layoutKeywords =
     ]
 
 data Symbol
-        = SymDot | SymSemi | SymVirtualSemi | SymBar
-        | SymColon | SymArrow | SymEqual | SymLambda
-        | SymUnderscore | SymQuestionMark   | SymAs
-        | SymOpenParen        | SymCloseParen
-        | SymOpenIdiomBracket | SymCloseIdiomBracket | SymEmptyIdiomBracket
-        | SymDoubleOpenBrace  | SymDoubleCloseBrace
-        | SymOpenBrace        | SymCloseBrace
-        | SymOpenVirtualBrace | SymCloseVirtualBrace
-        | SymOpenPragma       | SymClosePragma | SymEllipsis | SymDotDot
-        | SymEndComment -- ^ A misplaced end-comment "-}".
-    deriving (Eq, Show)
+  = SymDot          -- ^ @.@
+  | SymSemi         -- ^ @;@
+  | SymVirtualSemi  -- ^ layout-implied @;@
+  | SymBar          -- ^ @|@
+  | SymColon        -- ^ @:@
+  | SymArrow        -- ^ @->@ or @→@
+  | SymEqual        -- ^ @=@
+  | SymLambda       -- ^ @\@ or @λ@
+  | SymUnderscore   -- ^ @_@
+  | SymQuestionMark -- ^ @?@
+  | SymAs           -- ^ @@@
+  | SymOpenParen    -- ^ @(@
+  | SymCloseParen   -- ^ @)@
+
+  | SymEmptyIdiomBracket      -- ^ @(|)@ or @⦇⦈@
+  | SymOpenIdiomBracket  Bool -- ^ @(|@ or @⦇@, 'True' if unicode
+  | SymCloseIdiomBracket Bool -- ^ @|)@ or @⦈@, 'True' if unicode
+
+  | SymDoubleOpenBrace        -- ^ @{{@ or @⦃@
+  | SymDoubleCloseBrace       -- ^ @{{@ or @⦄@
+  | SymOpenBrace              -- ^ @{@
+  | SymCloseBrace             -- ^ @}@
+  | SymOpenVirtualBrace       -- ^ layout-implied @{@
+  | SymCloseVirtualBrace      -- ^ layout-implied @}@
+  | SymOpenPragma             -- ^ @{-#@
+  | SymClosePragma            -- ^ @#-}@
+  | SymEllipsis               -- ^ @...@ or @…@
+  | SymDotDot                 -- ^ @..@
+  | SymEndComment -- ^ A misplaced end-comment "-}".
+  deriving (Eq, Show)
+
+-- | "Tokens" which may appear qualified.
+data QualifiableToken
+  = QualDo              -- ^ qualified @do@
+  | QualEmptyIdiom      -- ^ qualified @(|)@ or @⦇⦈@
+  | QualOpenIdiom  Bool -- ^ qualified @(|@ or @⦇@, 'True' if unicode
+  deriving (Eq, Show)
+
+-- | A "qualified token".
+data QualifiedToken
+  = QualifiedToken
+      QualifiableToken     -- ^ The token.
+      [(Interval, String)] -- ^ Parts for the identifier
+      Interval             -- ^ Range of the @do@ keyword.
+  deriving (Eq, Show)
 
 data Token
           -- Keywords
@@ -87,6 +122,7 @@ data Token
         | TokQId        [(Interval, String)]
                         -- Non-empty namespace. The intervals for
                         -- "A.B.x" correspond to "A.", "B." and "x".
+        | TokQual_      QualifiedToken
           -- Literals
         | TokLiteral    RLiteral
           -- Special symbols
@@ -102,10 +138,19 @@ data Token
         | TokEOF Interval
     deriving (Eq, Show)
 
+pattern TokQual :: QualifiableToken -> [(Interval, String)] -> Interval -> Token
+pattern TokQual clz iss j = TokQual_ (QualifiedToken clz iss j)
+
+{-# COMPLETE TokKeyword, TokId, TokQId, TokQual, TokLiteral, TokSymbol, TokString, TokTeX, TokMarkup, TokComment, TokDummy, TokEOF #-}
+
+instance HasRange QualifiedToken where
+  getRange (QualifiedToken _ iss j) = getRange (map fst iss, j)
+
 instance HasRange Token where
   getRange (TokKeyword _ i)    = getRange i
   getRange (TokId (i, _))      = getRange i
   getRange (TokQId iss)        = getRange (map fst iss)
+  getRange (TokQual_ q)        = getRange q
   getRange (TokLiteral lit)    = getRange lit
   getRange (TokSymbol _ i)     = getRange i
   getRange (TokString (i, _))  = getRange i